requirements.txt
argon2-cffi==16.3.0
astroid==1.5.3
autoflake==0.7
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
coverage==4.4.1
cycler==0.10.0
Django==1.8
django-bootstrap3==9.0.0
django-braces==1.11.0
django-chartit==0.2.9
django-graphos==0.3.41
-e git+https://github.com/novapost/django-highcharts.git@34dd1b644325aaf8d9772feaf09827c730e8b154#egg=django_highcharts
django-jquery==3.1.0
django-markup-deprecated==0.0.3
django-rest-framework-multi-slug-field==0.2.0
django-tables2==1.10.0
djangorestframework==3.6.4
docutils==0.11
flake8==3.4.1
geoip2==2.5.0
idna==2.6
isort==4.2.15
Jinja2==2.9.6
lazy-object-proxy==1.3.1
Markdown==2.3.1
MarkupSafe==1.0
matplotlib==1.4.3
maxminddb==1.3.0
mccabe==0.6.1
mysqlclient==1.3.12
nose==1.3.7
numpy==1.10.1
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pylibmc==1.5.2
pylint==1.7.2
pymongo==2.5.1
pyparsing==2.0.3
python-dateutil==2.4.2
python3-memcached==1.51
pytz==2017.2
PyYAML==3.12
redis==2.7.4
requests==1.2.0
selenium==3.5.0
six==1.10.0
Sphinx==1.2b1
sqlparse==0.2.3
tblib==1.3.2
tornado==3.1
urllib3==1.22
wrapt==1.10.11
gunicorn>=19.7
whitenoise>=3.3.0
